Move Zarr/N5/neuroglancer test writers to data_fixtures.py; restrict to local paths
- Move write_zarr_v2_array, write_zarr_v3_array, write_n5_array,
  write_neuroglancer_precomputed_scale, write_ome_zarr_v2_group and
  write_ome_zarr_v3_group from tests/utils.py to tests/data_fixtures.py,
  alongside the other synthetic test-data writers (create_synthetic_multi_
  timepoint_ims), and update the importing test files accordingly.

- Restrict the new Zarr/N5/neuroglancer-precomputed converters to local
  input paths for now: TensorStoreChunkedImageSource.__init__ rejects a
  remote path (s3://, gs://, http(s)://, or any other non-local UPath) with
  the same "must be a local file path" ValueError MrcImageSource/
  ImsImageSource/CziImageSource already raise for their own formats, applied
  once in the shared base rather than duplicated per subclass. Updated the
  Changelog entry accordingly and added a rejection test per reader.
