+++ webknossos/Changelog.md
@@ -21,7 +21,7 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 - Removed the `timepoint` argument of `Dataset.add_layer_from_images`. Timeseries are now converted into a single layer with a `t` axis covering all timepoints, so selecting one only discarded data. Remove the argument from calls, and slice the `t` axis afterwards if you need a single timepoint. [#1477](https://github.com/scalableminds/webknossos-libs/pull/1477)
 
 ### Added
-- Added support for converting plain Zarr arrays (v2 and v3), OME-Zarr multiscale groups (NGFF 0.4 and 0.5), plain N5 datasets and N5 multiscale pyramids, and Neuroglancer precomputed volumes via `Dataset.from_images`/`add_layer_from_images`. For a multiscale source, the finest resolution level is converted as `Mag(1)`; the rest of the mag pyramid is built by the usual `downsample=True`/`Layer.downsample()` afterwards, as with every other format. Unlike `.ims`, MRC and CZI, these formats also support remote input paths (`s3://`, `gs://`, `http(s)://`). Builds on the `ChunkedImageSource` abstraction introduced for `.ims` and MRC. [#1495](https://github.com/scalableminds/webknossos-libs/pull/1495)
+- Added support for converting plain Zarr arrays (v2 and v3), OME-Zarr multiscale groups (NGFF 0.4 and 0.5), plain N5 datasets and N5 multiscale pyramids, and Neuroglancer precomputed volumes via `Dataset.from_images`/`add_layer_from_images`. For a multiscale source, the finest resolution level is converted as `Mag(1)`; the rest of the mag pyramid is built by the usual `downsample=True`/`Layer.downsample()` afterwards, as with every other format. Only local input paths are supported for now. Builds on the `ChunkedImageSource` abstraction introduced for `.ims` and MRC. [#1495](https://github.com/scalableminds/webknossos-libs/pull/1495)
 - Added support for converting Imaris (`.ims`) files via `Dataset.from_images`/`add_layer_from_images`, including multi-channel and multi-timepoint files. Converting a file that is both multi-channel and multi-timepoint yields one layer per channel, each keeping all of its timepoints. [#1477](https://github.com/scalableminds/webknossos-libs/pull/1477)
 - Added catchable exceptions for the ways image conversion can fail because of its input, so that callers of `Dataset.from_images`/`Dataset.add_layer_from_images` no longer have to match on error messages. All of them subclass the new `webknossos.ImageConversionError`, which subclasses `ValueError` (what these call sites raised before), and carry the offending `path`. [#1496](https://github.com/scalableminds/webknossos-libs/pull/1496)
     - `UnsupportedImageFormatError`: no reader handles this format. Also carries the `file_extension`, the `supported_file_extensions`, and `missing_extras` — the extras to install when the format is only unsupported because an optional dependency is missing, e.g. `.ims` without `webknossos[ims]`.
