+++ webknossos/tests/dataset/test_add_layer_from_images.py
@@ -24,16 +24,16 @@ from tests.constants import TESTDATA_DIR
 from tests.data_fixtures import (
     create_synthetic_multi_timepoint_ims,
     download_wklibs_sample_archive,
+    write_n5_array,
+    write_neuroglancer_precomputed_scale,
+    write_ome_zarr_v3_group,
+    write_zarr_v3_array,
 )
 from tests.utils import (
     HAS_PYLIBCZIRW,
     PYLIBCZIRW_EXPECTED,
     create_synthetic_czi,
     requires_pylibczirw,
-    write_n5_array,
-    write_neuroglancer_precomputed_scale,
-    write_ome_zarr_v3_group,
-    write_zarr_v3_array,
 )
